+++ cmake/developer_package/plugins/plugins.cmake
@@ -115,17 +115,17 @@ function(ie_add_plugin)
         # fake dependencies to build in the following order:
         # IE -> IE readers -> IE inference plugins -> IE-based apps
         if(BUILD_SHARED_LIBS)
-            if(TARGET ir_ngraph_frontend)
-                add_dependencies(${IE_PLUGIN_NAME} ir_ngraph_frontend)
+            if(TARGET ir_ov_frontend)
+                add_dependencies(${IE_PLUGIN_NAME} ir_ov_frontend)
             endif()
             if(TARGET inference_engine_ir_v7_reader)
                 add_dependencies(${IE_PLUGIN_NAME} inference_engine_ir_v7_reader)
             endif()
-            if(TARGET onnx_ngraph_frontend)
-                add_dependencies(${IE_PLUGIN_NAME} onnx_ngraph_frontend)
+            if(TARGET onnx_ov_frontend)
+                add_dependencies(${IE_PLUGIN_NAME} onnx_ov_frontend)
             endif()
-            if(TARGET paddlepaddle_ngraph_frontend)
-                add_dependencies(${IE_PLUGIN_NAME} paddlepaddle_ngraph_frontend)
+            if(TARGET paddlepaddle_ov_frontend)
+                add_dependencies(${IE_PLUGIN_NAME} paddlepaddle_ov_frontend)
             endif()
         endif()
