+++ cmake/templates/ngraphConfig.cmake.in
@@ -28,11 +28,11 @@
 #
 #   ngraph::frontend_manager               - nGraph frontend manager
 #
-#   ngraph_onnx_ngraph_frontend_FOUND      - True if the system has onnx_ngraph_frontend library
-#   ngraph::onnx_ngraph_frontend           - ONNX FrontEnd target (optional)
+#   ngraph_onnx_ov_frontend_FOUND          - True if the system has onnx_ov_frontend library
+#   ngraph::onnx_ov_frontend               - ONNX FrontEnd target (optional)
 #
 #   ngraph_paddlepaddle_frontend_FOUND     - True if the system has PDPD frontend
-#   ngraph::paddlepaddle_ngraph_frontend   - nGraph PDPD frontend (optional)
+#   ngraph::paddlepaddle_ov_frontend       - nGraph PDPD frontend (optional)
 #
 
 @PACKAGE_INIT@
@@ -58,38 +58,38 @@ if(TARGET openvino::frontend::manager AND NOT TARGET ngraph::frontend_manager)
         INTERFACE_LINK_LIBRARIES openvino::frontend::manager)
 endif()
 
-if(TARGET openvino::frontend::onnx AND NOT TARGET ngraph::onnx_ngraph_frontend)
-    add_library(ngraph::onnx_ngraph_frontend INTERFACE IMPORTED)
-    set_target_properties(ngraph::onnx_ngraph_frontend PROPERTIES
+if(TARGET openvino::frontend::onnx AND NOT TARGET ngraph::onnx_ov_frontend)
+    add_library(ngraph::onnx_ov_frontend INTERFACE IMPORTED)
+    set_target_properties(ngraph::onnx_ov_frontend PROPERTIES
         INTERFACE_LINK_LIBRARIES openvino::frontend::onnx)
 endif()
 
-if(TARGET openvino::frontend::paddlepaddle AND NOT TARGET ngraph::paddlepaddle_ngraph_frontend)
-    add_library(ngraph::paddlepaddle_ngraph_frontend INTERFACE IMPORTED)
-    set_target_properties(ngraph::paddlepaddle_ngraph_frontend PROPERTIES
+if(TARGET openvino::frontend::paddlepaddle AND NOT TARGET ngraph::paddlepaddle_ov_frontend)
+    add_library(ngraph::paddlepaddle_ov_frontend INTERFACE IMPORTED)
+    set_target_properties(ngraph::paddlepaddle_ov_frontend PROPERTIES
         INTERFACE_LINK_LIBRARIES openvino::frontend::paddlepaddle)
 endif()
 
-if(TARGET openvino::frontend::tensorflow AND NOT TARGET ngraph::tensorflow_ngraph_frontend)
-    add_library(ngraph::tensorflow_ngraph_frontend INTERFACE IMPORTED)
-    set_target_properties(ngraph::tensorflow_ngraph_frontend PROPERTIES
+if(TARGET openvino::frontend::tensorflow AND NOT TARGET ngraph::tensorflow_ov_frontend)
+    add_library(ngraph::tensorflow_ov_frontend INTERFACE IMPORTED)
+    set_target_properties(ngraph::tensorflow_ov_frontend PROPERTIES
         INTERFACE_LINK_LIBRARIES openvino::frontend::tensorflow)
 endif()
 
 set(ngraph_ngraph_FOUND ON)
 set(NGRAPH_LIBRARIES ngraph::ngraph)
 
-set(ngraph_onnx_ngraph_frontend_FOUND ${OpenVINO_Frontend_ONNX_FOUND})
+set(ngraph_onnx_ov_frontend_FOUND ${OpenVINO_Frontend_ONNX_FOUND})
 set(ngraph_onnx_importer_FOUND ${OpenVINO_Frontend_ONNX_FOUND})
 
 if(ngraph_onnx_importer_FOUND)
-    set(ONNX_IMPORTER_LIBRARIES ngraph::onnx_ngraph_frontend)
+    set(ONNX_IMPORTER_LIBRARIES ngraph::onnx_ov_frontend)
     # ngraph::onnx_importer target and variables are deprecated
     # but need to create a dummy target for BW compatibility
     if(NOT TARGET ngraph::onnx_importer)
         add_library(ngraph::onnx_importer INTERFACE IMPORTED)
         set_target_properties(ngraph::onnx_importer PROPERTIES
-            INTERFACE_LINK_LIBRARIES ngraph::onnx_ngraph_frontend)
+            INTERFACE_LINK_LIBRARIES ngraph::onnx_ov_frontend)
     endif()
 endif()
