+++ docs/IE_DG/Deep_Learning_Inference_Engine_DevGuide.md
@@ -40,13 +40,13 @@ This library contains the classes to:
 
 Starting from 2022.1 release, OpenVINO Runtime introduced a concept of frontend plugins. Such plugins can be automatically dynamically loaded by OpenVINO Runtime dynamically depending on file format:
 * Linux* OS:
-    - `libir_ngraph_frontend.so` to read a network from IR
-    - `libpaddlepaddle_ngraph_frontend.so` to read a network from PaddlePaddle model format
-    - `libonnx_ngraph_frontend.so` to read a network from ONNX model format
+    - `libir_ov_frontend.so` to read a network from IR
+    - `libpaddlepaddle_ov_frontend.so` to read a network from PaddlePaddle model format
+    - `libonnx_ov_frontend.so` to read a network from ONNX model format
 * Windows* OS:
-    - `ir_ngraph_frontend.dll` to read a network from IR
-    - `paddlepaddle_ngraph_frontend.dll` to read a network from PaddlePaddle model format
-    - `onnx_ngraph_frontend.dll` to read a network from ONNX model format
+    - `ir_ov_frontend.dll` to read a network from IR
+    - `paddlepaddle_ov_frontend.dll` to read a network from PaddlePaddle model format
+    - `onnx_ov_frontend.dll` to read a network from ONNX model format
 
 ### Device-Specific Plugin Libraries
