+++ docs/snippets/CMakeLists.txt
@@ -46,7 +46,7 @@ if(OpenCV_FOUND)
 endif()
 
 if(NGRAPH_ONNX_FRONTEND_ENABLE)
-    target_link_libraries(${TARGET_NAME} PRIVATE onnx_ngraph_frontend)
+    target_link_libraries(${TARGET_NAME} PRIVATE onnx_ov_frontend)
 endif()
 
 if(NOT MSVC)
