+++ inference-engine/src/CMakeLists.txt
@@ -60,20 +60,20 @@ endif()
 
 if(NGRAPH_IR_FRONTEND_ENABLE)
     if(BUILD_SHARED_LIBS)
-        add_dependencies(ie_libraries ir_ngraph_frontend)
+        add_dependencies(ie_libraries ir_ov_frontend)
     endif()
     # use this one once CVS-69781 is fixed
-    # add_dependencies(inference_engine ir_ngraph_frontend)
+    # add_dependencies(inference_engine ir_ov_frontend)
 endif()
 
 if(NGRAPH_ONNX_FRONTEND_ENABLE)
-    add_dependencies(inference_engine onnx_ngraph_frontend)
+    add_dependencies(inference_engine onnx_ov_frontend)
 endif()
 
 if(NGRAPH_PDPD_FRONTEND_ENABLE)
-    add_dependencies(inference_engine paddlepaddle_ngraph_frontend)
+    add_dependencies(inference_engine paddlepaddle_ov_frontend)
 endif()
 
 if(NGRAPH_TF_FRONTEND_ENABLE)
-    add_dependencies(inference_engine tensorflow_ngraph_frontend)
+    add_dependencies(inference_engine tensorflow_ov_frontend)
 endif()
