+++ inference-engine/tests/functional/plugin/shared/CMakeLists.txt
@@ -18,7 +18,7 @@ endif()
 
 # remove once CVS-69781 is fixed
 if(NGRAPH_IR_FRONTEND_ENABLE)
-    list(APPEND DEPENDENCIES ir_ngraph_frontend)
+    list(APPEND DEPENDENCIES ir_ov_frontend)
 endif()
 
 if (NGRAPH_ONNX_FRONTEND_ENABLE)
