+++ inference-engine/tests/unit/frontends/onnx_import/CMakeLists.txt
@@ -11,7 +11,7 @@ addIeTargetTest(
             gtest
             gtest_main
             commonTestUtils
-            onnx_ngraph_frontend
+            onnx_ov_frontend
             openvino::util
         DEFINES
             ONNX_MODELS_DIR=\"${TEST_MODEL_ZOO}/onnx_import\"
