+++ model-optimizer/unit_tests/mock_mo_frontend/mock_mo_ov_frontend/mock_mo_frontend.cpp
@@ -29,7 +29,7 @@ extern "C" MOCK_API FrontEndVersion GetAPIVersion()
 extern "C" MOCK_API void* GetFrontEndData()
 {
     FrontEndPluginInfo* res = new FrontEndPluginInfo();
-    res->m_name = "mock_mo_ngraph_frontend";
+    res->m_name = "mock_mo_ov_frontend";
     res->m_creator = []() { return std::make_shared<FrontEndMockPy>(); };
 
     return res;
