+++ model-optimizer/unit_tests/mock_mo_frontend/mock_mo_python_api/CMakeLists.txt
@@ -21,7 +21,7 @@ set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${PYTHON_BRIDGE_OUTPUT_DIRECTORY})
 set(CMAKE_COMPILE_PDB_OUTPUT_DIRECTORY ${PYTHON_BRIDGE_OUTPUT_DIRECTORY})
 set(CMAKE_PDB_OUTPUT_DIRECTORY ${PYTHON_BRIDGE_OUTPUT_DIRECTORY})
 
-set(TARGET_FE_NAME "mock_mo_ngraph_frontend")
+set(TARGET_FE_NAME "mock_mo_ov_frontend")
 set(PYBIND_FE_NAME "mock_mo_python_api")
 
 set(PYBIND_FE_SRC ${CMAKE_CURRENT_SOURCE_DIR}/mock_mo_python_api.cpp)
