+++ model-optimizer/unit_tests/mock_mo_frontend/mock_mo_python_api/mock_mo_python_api.cpp
@@ -5,7 +5,7 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-#include "../mock_mo_ngraph_frontend/mock_mo_frontend.hpp"
+#include "../mock_mo_ov_frontend/mock_mo_frontend.hpp"
 
 namespace py = pybind11;
 using namespace ngraph;
