+++ ngraph/frontend/CMakeLists.txt
@@ -3,6 +3,7 @@
 #
 
 set(FRONTEND_INSTALL_INCLUDE "runtime/include/ngraph/frontend")
+set(FRONTEND_NAME_SUFFIX "_ov_frontend")
 
 add_subdirectory(frontend_manager)
