+++ ngraph/frontend/frontend_manager/CMakeLists.txt
@@ -33,6 +33,11 @@ target_include_directories(${TARGET_NAME} PUBLIC
         $<INSTALL_INTERFACE:${FRONTEND_INSTALL_INCLUDE}>)
 target_include_directories(${TARGET_NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/src)
 target_link_libraries(${TARGET_NAME} PRIVATE ${CMAKE_DL_LIBS} openvino::util PUBLIC ngraph)
+target_compile_definitions(${TARGET_NAME}
+        PRIVATE
+        FRONTEND_LIB_PREFIX="${CMAKE_SHARED_LIBRARY_PREFIX}"
+        FRONTEND_LIB_SUFFIX="${FRONTEND_NAME_SUFFIX}${IE_BUILD_POSTFIX}"
+        )
 
 add_clang_format_target(${TARGET_NAME}_clang FOR_TARGETS ${TARGET_NAME})
 
@@ -56,6 +61,11 @@ if(BUILD_SHARED_LIBS)
         $<BUILD_INTERFACE:${FRONTEND_INCLUDE_DIR}>)
     target_include_directories(${TARGET_NAME}_static PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/src)
     target_compile_definitions(${TARGET_NAME}_static PUBLIC USE_STATIC_FRONTEND_MANAGER)
+    target_compile_definitions(${TARGET_NAME}_static
+            PRIVATE
+            FRONTEND_LIB_PREFIX="${CMAKE_SHARED_LIBRARY_PREFIX}"
+            FRONTEND_LIB_SUFFIX="${FRONTEND_NAME_SUFFIX}${IE_BUILD_POSTFIX}"
+            )
     ov_install_static_lib(${TARGET_NAME}_static ngraph)
 else()
     add_library(${TARGET_NAME}::static ALIAS ${TARGET_NAME})
