+++ ngraph/frontend/onnx/frontend/include/onnx_import/onnx_importer_visibility.hpp
@@ -4,7 +4,7 @@
 
 #include "ngraph/visibility.hpp"
 
-#ifdef onnx_ngraph_frontend_EXPORTS
+#ifdef onnx_ov_frontend_EXPORTS
 #    define ONNX_IMPORTER_API OPENVINO_CORE_EXPORTS
 #else
 #    define ONNX_IMPORTER_API OPENVINO_CORE_IMPORTS
