+++ ngraph/frontend/paddlepaddle/CMakeLists.txt
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-set(TARGET_NAME "paddlepaddle_ngraph_frontend")
+set(TARGET_NAME "paddlepaddle${FRONTEND_NAME_SUFFIX}")
 
 file(GLOB_RECURSE LIBRARY_SRC ${CMAKE_CURRENT_SOURCE_DIR}/src/*.cpp)
 file(GLOB_RECURSE LIBRARY_HEADERS ${CMAKE_CURRENT_SOURCE_DIR}/src/*.hpp)
@@ -51,7 +51,7 @@ set_source_files_properties(${PROTO_SRCS} ${PROTO_HDRS} PROPERTIES COMPILE_OPTIO
 add_library(${TARGET_NAME} SHARED ${LIBRARY_SRC} ${LIBRARY_HEADERS} ${LIBRARY_PUBLIC_HEADERS} ${PROTO_SRCS} ${PROTO_HDRS})
 add_library(openvino::frontend::paddlepaddle ALIAS ${TARGET_NAME})
 
-add_dependencies(${TARGET_NAME} paddlepaddle_ngraph_frontend_proto)
+add_dependencies(${TARGET_NAME} ${TARGET_NAME}_proto)
 
 ov_ncc_naming_style(FOR_TARGET ${TARGET_NAME}
                     INCLUDE_DIRECTORY "${CMAKE_CURRENT_SOURCE_DIR}/include"
