+++ ngraph/frontend/paddlepaddle/include/paddlepaddle_frontend/utility.hpp
@@ -7,11 +7,11 @@
 #include <frontend_manager/frontend_manager.hpp>
 
 // Defined if we are building the plugin DLL (instead of using it)
-#ifdef paddlepaddle_ngraph_frontend_EXPORTS
+#ifdef paddlepaddle_ov_frontend_EXPORTS
 #    define PDPD_API OPENVINO_CORE_EXPORTS
 #else
 #    define PDPD_API OPENVINO_CORE_IMPORTS
-#endif  // paddlepaddle_ngraph_frontend_EXPORTS
+#endif  // paddlepaddle_ov_frontend_EXPORTS
 
 #define PDPD_ASSERT(ex, msg)               \
     {                                      \
