+++ ngraph/frontend/tensorflow/include/tensorflow_frontend/utility.hpp
@@ -6,8 +6,8 @@
 
 #include "frontend_manager/frontend_exceptions.hpp"
 
-#ifdef tensorflow_ngraph_frontend_EXPORTS
+#ifdef tensorflow_ov_frontend_EXPORTS
 #    define TF_API OPENVINO_CORE_EXPORTS
 #else
 #    define TF_API OPENVINO_CORE_IMPORTS
-#endif  // tensorflow_ngraph_frontend_EXPORTS
+#endif  // tensorflow_ov_frontend_EXPORTS
