+++ ngraph/test/CMakeLists.txt
@@ -614,9 +614,9 @@ if (OV_COMPILER_IS_CLANG)
 endif()
 
 if (NGRAPH_ONNX_FRONTEND_ENABLE)
-    get_target_property(ONNX_FRONTEND_SRC_DIR onnx_ngraph_frontend SOURCE_DIR)
+    get_target_property(ONNX_FRONTEND_SRC_DIR onnx_ov_frontend SOURCE_DIR)
     target_include_directories(unit-test PRIVATE ${ONNX_FRONTEND_SRC_DIR}/src)
-    target_link_libraries(unit-test PRIVATE onnx_ngraph_frontend onnx_test_util)
+    target_link_libraries(unit-test PRIVATE onnx_ov_frontend onnx_test_util)
     if (LINUX)
         target_link_options(unit-test PRIVATE -Wl,--exclude-libs,ALL)
     elseif(APPLE)
@@ -625,7 +625,7 @@ if (NGRAPH_ONNX_FRONTEND_ENABLE)
 endif()
 
 if(NGRAPH_IR_FRONTEND_ENABLE)
-    add_dependencies(unit-test ir_ngraph_frontend)
+    add_dependencies(unit-test ir_ov_frontend)
 endif()
 
 install(TARGETS unit-test
