+++ ngraph/test/frontend/onnx/CMakeLists.txt
@@ -22,5 +22,5 @@ target_compile_definitions(${TARGET_NAME} PRIVATE -D TEST_ONNX_MODELS_DIRNAME=\"
 set(MANIFEST ${CMAKE_CURRENT_SOURCE_DIR}/unit_test.manifest)
 target_compile_definitions(${TARGET_NAME} PRIVATE -D MANIFEST=\"${MANIFEST}\")
 
-add_dependencies(${TARGET_NAME} onnx_ngraph_frontend)
+add_dependencies(${TARGET_NAME} onnx_ov_frontend)
 add_dependencies(${TARGET_NAME} test_model_zoo)
