+++ runtime/bindings/python/CMakeLists.txt
@@ -17,9 +17,9 @@ set(LIBRARY_OUTPUT_DIRECTORY_BIN ${CMAKE_LIBRARY_OUTPUT_DIRECTORY})
 add_subdirectory(src)
 
 if(NGRAPH_UNIT_TEST_ENABLE)
-    add_subdirectory(tests/mock/mock_py_ngraph_frontend)
-    add_dependencies(_pyngraph mock_py_ngraph_frontend)
-    set_target_properties(mock_py_ngraph_frontend PROPERTIES
+    add_subdirectory(tests/mock/mock_py_ov_frontend)
+    add_dependencies(_pyngraph mock_py_ov_frontend)
+    set_target_properties(mock_py_ov_frontend PROPERTIES
             LIBRARY_OUTPUT_DIRECTORY ${LIBRARY_OUTPUT_DIRECTORY_BIN}
             ARCHIVE_OUTPUT_DIRECTORY ${LIBRARY_OUTPUT_DIRECTORY_BIN}
             COMPILE_PDB_OUTPUT_DIRECTORY ${LIBRARY_OUTPUT_DIRECTORY_BIN}
