+++ runtime/bindings/python/tests_compatibility/mock/pyngraph_fe_mock_api/pyngraph_mock_frontend_api.cpp
@@ -5,7 +5,7 @@
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
 
-#include "../mock_py_ngraph_frontend/mock_py_frontend.hpp"
+#include "../mock_py_ov_frontend/mock_py_frontend.hpp"
 
 namespace py = pybind11;
 using namespace ngraph;
